LPS-60606 Move StagingPortletConfigurationIcon classes to StagingConfigurationWeb and removed addon since it's no longer used
