+++ modules/apps/staging/staging-configuration-web/src/main/java/com/liferay/staging/configuration/web/configuration/icon/StagingPortletConfigurationIconFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.configuration.icon.staging;
+package com.liferay.staging.configuration.web.configuration.icon;
 
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIconFactory;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
